return Iterators of Results instead of Vecs
Return Results of Iterators of Results instead of Results of Vecs. Is it
harder to read with a bunch of .maps, .and_thens, and closures? Yes. Is
it more precise and potentially more tolerant to one website erroring?
Yes. Does it keep us from having to store potentially 100+ DOM trees in
memory at once in the future? Also yes. Is it worth it? Maybe, maybe
not. But I can always revert it if it's not a good idea.