fix traversal_parent
Fix a panic in `traversal_parent`. If `node.parent()` returns
`Node::Document`, we are at the root of the tree so gracefully return
`None` to signal the end. If it returns a `Node::Element(e)` then wrap
it up in an `ElementRef` (without checking again whether it's an
`Element`). If it is any other kind of `Node`, `panic!()` because I
don't know how to handle it.

This stops github benchmark from panicking.