create and implement node_id debugging function
Declare `SelectorMapElement::node_id` debugging function in stylo, and
implement it in mach_6. This computes the hash every time as opposed to
using a stored ID like in mach_6's Element wrapper type.

This is a solution to help us debug the algorithms inside Stylo using
print statements; Stylo only works with `SelectorMapElement`s, so we
need to be able to get each element's node id from that.